requirements.txt
selenium==4.41.0
pytest==9.0.2
allure-pytest==2.15.3
pytest-xdist==3.8.0
Faker==40.13.0